requirements.txt
asgiref==3.5.1
attrs==21.4.0
coverage==6.4
Django==3.2.13
django-cors-headers==3.12.0
django-nose==1.4.7
djangorestframework==3.13.1
iniconfig==1.1.1
nose==1.3.7
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
